refactor: Dashboard JS split: extract scoreboard.js (per-domain split, … [ORB-00173]
Dashboard JS split: extract scoreboard.js (per-domain split, round 4)
